🔥 fix : create and move to core directory
