+++ src/core/App.test.js
