+++ src/index.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import './index.css';
-import App from './App';
-import reportWebVitals from './reportWebVitals';
+//import './index.css';
+import App from './core/App';
+import reportWebVitals from './core/reportWebVitals';
 import { Provider } from 'react-redux';
 import store from 'src/store';  // Redux 스토어를 가져옵니다.
